Add P279 to PropertySuggesterClassifyingPropertyIds per default.
The default of PropertySuggesterClassifyingPropertyIds should be
the same as the default of the classifying_properties setting in
analyzer.ini in PropertySuggester-Python.

Note: as per T169058, this setting is set explicitly in production,
so this change to the default does not have any effect in production.
See T169060 for changing this setting in production.

Bug: T169057
Change-Id: Id2b1626a930e662fed334f9848bc6ceb6c912315