Add code required to carry out A/B testing with the two property suggesters
Added code requred to handle event logging both client side and server
side. When a request is made with the 'context' parameter 'item' an
event will be logged. These events will record which suggester was used
and additional performance metrics.

The ResourceLoader module is moved to the ResourceLoaderRegisterModules
hook so that it can have a conditional dependency on EventLogging.

Change-Id: Ia1170a183f1ada7447e030b1f97bac05c56a4508