Revert changes to `redirect` table after test runs
GetSuggestionsTest inserts data into the `redirect` table. In
versions of MediaWiki before 1.41 it was necessary to manually
specify which tables should be rolled back at the end of a test.
Add `redirect` to the list of tables used in the test to avoid
leaving data around that interferes with the execution of other
tests (specifically `MergeLexemesTest` and
`SpecialMergeLexemesTest`.

Bug: T360435
Change-Id: I7c8e59933f9a93b86268598885f1b29c0bf50ac1