Add support for stream paths (`php://stdin`) for CSV import
To support migration of the `UpdateTable` maintenance script from
`mwscript` to `mwscript-k8s`, we need to move away from passing
one very large file to the script as a path - the kubernetes
implementation of `mwscript` can only handle files of up to 1MiB.

Change the `UpdateTable` script to work with PHP streams (e.g.
`php://stdin`) so that we can stream the data to `mwscript-k8s`.

Bug: T376604
Change-Id: I0e799332f1e702358607f1a636c1f33e8e4142c7